# mib/Specification/swagger.yml
openapi: 3.0.1

info:
  title: Message Service API
  description: Message Microservice of Message in a Bottle Application
  version: 0.2-oas3

servers:
  - url: /

paths:
  /mailbox:
    post:
      tags: [ Message ]
      summary: show messages
      operationId: mib.resource.message.mailbox
      requestBody:
        description: show messages
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                id:
                  type: string
                filter:
                  type: string
      description: |
        messages info
      responses:
        201:
          description: you can't see this information
        202:
          description: messages retrived correctly
  /delete_message/{draft_id}:
    get:
      tags: [ Message ]
      description: delete drfat message by its id
      operationId: mib.resource.message.delete_message
      parameters:
        - name: draft_id
          description: Draft message Unique id
          in: path
          required: true
          schema:
            type: string
      responses:
        200:
          description: Delete done
  /draft_message_info/{draft_id}:
    get:
      tags: [ Message ]
      description: give drfat message info by its id
      operationId: mib.resource.message.draft_message_info
      parameters:
        - name: draft_id
          description: Draft message Unique id
          in: path
          required: true
          schema:
            type: string
      responses:
        200:
          description: draft message info
  /send_message:
    post:
      tags: [ Message ]
      summary: add message
      operationId: mib.resource.message.send_message
      requestBody:
        description: message sent
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - sender_id
                - sender_nickname
                - receiver_id
                - receiver_nickname
                - body
                - delivery_date
                - image
              properties:
                sender_id:
                  type: string
                sender_nickname:
                  type: string
                receiver_id:
                  type: string
                receiver_nickname:
                  type: string
                body:
                  type: string
                delivery_date:
                  type: string
                image:
                  type: string
      description: |
        add message
      responses:
        202:
          description: message sent
        303:
          description: message not sent
  /draft_message:
    post:
      tags: [ Message ]
      summary: draft message
      operationId: mib.resource.message.draft_message
      requestBody:
        description: message drafted
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - sender_id
                - sender_nickname
                - receiver_id
                - receiver_nickname
                - body
                - delivery_date
                - image
              properties:
                sender_id:
                  type: string
                sender_nickname:
                  type: string
                receiver_id:
                  type: string
                receiver_nickname:
                  type: string
                body:
                  type: string
                delivery_date:
                  type: string
                image:
                  type: string
      description: |
        draft message
      responses:
        202:
          description: message drafted
        303:
          description: message not drafted
  /send_draft_message:
    post:
      tags: [ Message ]
      summary: send draft message
      operationId: mib.resource.message.send_draft_message
      requestBody:
        description: send a draft message 
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - sender_id
                - sender_nickname
                - receiver_id
                - receiver_nickname
                - body
                - delivery_date
                - image
                - draft_id
              properties:
                sender_id:
                  type: string
                sender_nickname:
                  type: string
                receiver_id:
                  type: string
                receiver_nickname:
                  type: string
                body:
                  type: string
                delivery_date:
                  type: string
                image:
                  type: string
                draft_id:
                  type: string
      description: |
        send a draft message
      responses:
        202:
          description: draft message sent
        303:
          description: draft message not sent
  /update_draft_message:
    post:
      tags: [ Message ]
      summary: update draft message
      operationId: mib.resource.message.update_draft_message
      requestBody:
        description: update a draft message 
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - sender_id
                - sender_nickname
                - receiver_id
                - receiver_nickname
                - body
                - delivery_date
                - image
                - draft_id
              properties:
                sender_id:
                  type: string
                sender_nickname:
                  type: string
                receiver_id:
                  type: string
                receiver_nickname:
                  type: string
                body:
                  type: string
                delivery_date:
                  type: string
                image:
                  type: string
                draft_id:
                  type: string
      description: |
        update a draft message
      responses:
        202:
          description: draft message update
        303:
          description: draft message not update
  /delete_received_message/{id}:
    get:
      tags: [ Message ]
      description: delete received message by its id
      operationId: mib.resource.message.delete_received_message
      parameters:
        - name: id
          description: received message Unique id
          in: path
          required: true
          schema:
            type: string
      responses:
        200:
          description: Delete done
  /open_received_message/{id}:
    get:
      tags: [ Message ]
      description: opne received message by its id
      operationId: mib.resource.message.open_received_message
      parameters:
        - name: id
          description: open message Unique id
          in: path
          required: true
          schema:
            type: string
      responses:
        200:
          description: open done
  /open_send_message/{id}:
    get:
      tags: [ Message ]
      description: open send message by its id
      operationId: mib.resource.message.open_send_message
      parameters:
        - name: id
          description: open message Unique id
          in: path
          required: true
          schema:
            type: string
      responses:
        200:
          description: open done
  /calendar:
    post:
      tags: [ Message ]
      summary: show messages
      operationId: mib.resource.message.calendar
      requestBody:
        description: show messages
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                id:
                  type: string
                filter:
                  type: string
      description: |
        messages info
      responses:
        201:
          description: you can't see this information
        202:
          description: messages retrived correctly